# Adds automatic support for libdeflate which yields at least a 2x speedup