--- a/cdflib/gzip_wrapper.py
+++ b/cdflib/gzip_wrapper.py
@@ -1,0 +1,7 @@
+try:
+    from deflate import gzip_decompress as gzip_inflate
+    from deflate import gzip_compress as gzip_deflate
+
+except ImportError:
+    from gzip import decompress as gzip_inflate
+    from gzip import compress as gzip_deflate