This PR allows to use deflate package when available. It could be added later a either dependency or optional dependency.
Here are some speedup examples:

Without libdeflate:
```
In [12]: %timeit cdflib.CDF("/home/jeandet/Downloads/bigcdf_compressed.cdf").varget("temp")
5.35 s ± 33.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [13]: %timeit cdflib.CDF("/home/jeandet/cdf_benchmarks/solo_l3_rpw-bia-efield_20220906_v01.cdf").varget("EDC_SRF")
786 ms ± 5.26 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

With libdeflate:
```
In [8]: %timeit cdflib.CDF("/home/jeandet/Downloads/bigcdf_compressed.cdf").varget("temp")
1.19 s ± 23.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [9]: %timeit cdflib.CDF("/home/jeandet/cdf_benchmarks/solo_l3_rpw-bia-efield_20220906_v01.cdf").varget("EDC_SRF")
503 ms ± 5.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```