@dstansby 2023-07-28
Nice! I'm pro not having this as an optional dependency for simplicity, but looks like deflate isn't on conda-forge for now so best to leave it as optional for now.